refactor: 파일 분리, MARK 주석 추가
