+++ OpenMarket/OpenMarket/Extension/StringExtension.swift
@@ -1,0 +1,17 @@
+//
+//  StringExtension.swift
+//  OpenMarket
+//
+//  Created by kakao on 2022/01/27.
+//
+
+import UIKit
+extension String {
+    func strikeThrough() -> NSAttributedString {
+        let attributeString: NSMutableAttributedString = NSMutableAttributedString(string: self)
+        attributeString.addAttribute(NSAttributedString.Key.strikethroughStyle,
+                                     value: NSUnderlineStyle.single.rawValue,
+                                     range: NSRange(location: 0, length: attributeString.length))
+        return attributeString
+    }
+}
