+++ OpenMarket/OpenMarket/Extension/UIViewExtension.swift
@@ -1,0 +1,14 @@
+//
+//  UIViewExtension.swift
+//  OpenMarket
+//
+//  Created by kakao on 2022/01/27.
+//
+
+import UIKit.UIView
+
+extension UIView {
+    static var identifier: String {
+        String(describing: self)
+    }
+}
