+++ OpenMarket/OpenMarket/Resource/SceneDelegate.swift
@@ -18,7 +18,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
         let navigationController: UINavigationController = UINavigationController()
-        navigationController.pushViewController(ViewController(), animated: true)
+        navigationController.pushViewController(ProductListViewController(), animated: true)
         window?.rootViewController = navigationController
         window?.makeKeyAndVisible()
     }
